Move `logging` types to module
Move all `logging` types into the `logging` module.

This moves the repo from a `layer` architecture where files are
organised according to layer, to a domain model where files are
organised according to what they relate to.
